Update Swagger definitions
